requirements.txt
torch==1.13.1
torchvision==0.13.1
numpy >= 1.19
pyyaml~=6.0
tensorboard
tqdm~=4.64.1
git+https://github.com/Ofir408/BEHRT.git@master
pytorch-pretrained-bert
h5py